@@ -19,7 +19,7 @@ publish_to: "none"
 version: 1.0.0+1
 
 environment:
-    sdk: ">=2.17.5 <3.0.0"
+    sdk: ">=3.0.0 <4.0.0"
 
 dependencies:
     flutter:
@@ -30,7 +30,7 @@ dependencies:
     cupertino_icons: ^1.0.2
     font_awesome_flutter: ^10.1.0
     flutter_rating_bar: ^4.0.1
-    intl: ^0.17.0
+    intl: ^0.20.2
     animations: ^2.0.3
 
 dev_dependencies:
